side fixes: Moving authentication logic in a dedicated directory at service layer
